Move draftail out of components folder
