+++ client/src/draftail/blocks/EmbedBlock.js
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import { gettext } from '../../../utils/gettext';
-import MediaBlock from '../blocks/MediaBlock';
+import { gettext } from '../../utils/gettext';
+import MediaBlock from './MediaBlock';
 
 /**
  * Editor block to display media and edit content.
