+++ client/src/draftail/blocks/ImageBlock.scss
