+++ client/src/draftail/blocks/MediaBlock.scss
