+++ client/src/draftail/decorators/TooltipEntity.scss
