+++ client/src/draftail/decorators/TooltipEntity.test.js
