+++ wagtail/admin/static_src/wagtailadmin/scss/panels/draftail.scss
@@ -1,3 +1,3 @@
 @import '../../../../../../client/scss/settings';
 @import '../../../../../../client/scss/tools';
-@import '../../../../../../client/src/components/Draftail/Draftail';
+@import '../../../../../../client/src/draftail/Draftail';
